use reportConfig global variable

+++ grafana_collector/grafana/time.go
@@ -206,7 +206,7 @@ func (n now) parseRelativeTime(s string) time.Time {
 
 func parseAbsTime(s string) time.Time {
 	if timeInMs, err := strconv.ParseInt(s, 10, 64); err == nil {
-		return time.Unix(int64(timeInMs)/1000, 0)
+		return time.Unix(timeInMs/1000, 0)
 	}
 
 	panic(unrecognized(s))
